requirements.txt
torch
llama-cpp-python
mss
Pillow
pyautogui
python-dotenv